Add independent server interface
In order to not depend on any WebSocket library, there should be an
independent server interface to handle a variety of libaries. A single,
high-level interface improves consistency and eliminates the need to
depend on a single library. In addition, it also adds flexibility to add
more features if necessary.

The interface includes the following operations:
- Sending packet data to client
- Receiving packet data from client distinguished with the appropriate
  header attribute
- Selector interface to select clients with a predicate or a collection