Fix identifiers of parameters in server interface
Some parameters had the name "raw", indicating that it is a union of
types rather than a singular type. Even though it is great to distinguish
the parameter types in implementation level, it is not appriopiate for the
interface and should contain a convenient identifier as if the user knew
nothing about the implementation details at all.