Enhance watcher with client/server synchronization
The gulp task runner now use nodemon to restart the server on file
change. After running the server, browser sync will be called to
automatically refresh the page to stay synchronized.

This cycle greatly improves productivity as the tedious work such as
running the script and opening the browser completely got eliminated.