Add basic object serializer and type validator
As many operations will not be send frequently and are not too large,
there is no really need for advanced compression strategy and smart binary
manipulation. A simple BSON serializer will fit for such packets.

The interface is a wrapper around the BSON serializer. It also includes a
simple type validator, to make sure that the packet is valid. It is also a
consequence for the fact that the client packets never should be trusted.