Fix send function in socket interface
Formerly the send function in socket interface only expected binary data
as input. But it knew nothing about the header information whatsoever,
causing every message that does not include a header be sended to
system(0x00). This behaviour is of course unacceptable.

This bug also has been fixed for bpio interface.