Add client 2d graphics interface
Many simple draw operations in the build-in canvas API are mostly tedious
to call and overly complicated. The canvas API also does not support a
way to draw shapes independent of the user window size and aspect ratio.

All these issues are fixed by introducing a wrapper around the canvas
API with a simple to use interface. In addition, it will add a layer of
abstraction where there is a possibility to perform all sorts of
optimization, e.g call batching and caching draw operations.