Improve client socket and implement client engine
The client socket interface used to not perform properly on connection
lost. The whole socket object had to be instantiated to reconnect to the
server. Now the interface includes a separate attribute -- named system --
where sockets can be reconnected and querying for connection status on
connection lost.

There is also an implementation for abstract scene class and client engine
manager. Handy features such as, input querying and scene switching are
all implemented. Also the socket gets automatically reconnected on
connection lost.